requirements.txt
flask==3.0.3
Werkzeug==3.0.4
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.3
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.2.0
SQLAlchemy>=2.0.29
alembic>=1.13.1
python-dotenv>=1.0.1
jinja2>=3.1.4
pytest>=8.2.0
requests>=2.32.0
git-lfs